RGW: Update structure and fix errors
